Add files via upload
here we are using int data type and checking its type or class and converting it into string and float
overview

Int To String  ✔
Float To String ✔
Complex To String ✔

String To Int   ❌
Int To Float ✔
Complex To Int ❌

String To Float  ❌   ✔ [if we take "13(number)" it convert. but  in string "name" will not convert]
Int To Float  ✔
Complex To Float  ❌ 

Int To Complex  ✔
String To Complex  ✔  ❌   [ if "number" yes. if "string" no]
Float To Complex ✔